Split simulator_handle.cpp into a source and header file
