+++ src/io/simulator/CMakeLists.txt
@@ -1,8 +1,5 @@
 set(io_simulator_sources
-    simulator.hpp
-    simulator_handle.hpp
-    simulator_stats.hpp
-    simulator_config.hpp)
+    simulator_handle.cpp)
 
 find_package(fmt REQUIRED)
 find_package(Threads REQUIRED)
